getting test environments working
