+++ package.json
@@ -12,6 +12,13 @@
   "bugs": {
     "url": "https://github.com/ipfs-shipyard/js-pinning-service-http-client/issues"
   },
+  "keywords": [
+    "dweb",
+    "ipfs",
+    "pinning service",
+    "protocol labs",
+    "web3"
+  ],
   "engines": {
     "node": ">=16.0.0",
     "npm": ">=7.0.0"
@@ -126,14 +133,17 @@
   "scripts": {
     "dep-check": "aegir dep-check src/**/*.ts test/**/*.ts generated/**/*.ts",
     "fix": "run-s fix:*",
-    "fix:gen": "./npm-scripts/fix/generated-code.sh",
+    "fix:gen": "./npm-scripts/fix/gen.sh",
     "fix:lint": "aegir lint --fix",
     "lint": "run-s lint:*",
     "lint:main": "aegir lint",
-    "lint:project": "check-aegir-project",
+    "lint:ts": "aegir ts -p check -- -p tsconfig.json --suppress 6133@generated 6192@generated --stats",
+    "lint-TODO:project": "check-aegir-project # currently broken due to corrupting the repoUrl",
     "release": "aegir release",
     "postinstall": "./npm-scripts/postinstall.sh",
     "build": "aegir build -- -p tsconfig.json --suppress 6133@generated 6192@generated --stats",
+    "build:main": "aegir build -- -p tsconfig.json --suppress 6133@generated 6192@generated --stats",
+    "build:docs": "aegir ts docs",
     "test": "run-p test:*",
     "test:electron": "aegir test --target electron-main",
     "test:node": "aegir test --target node --cov && npx nyc report",
@@ -150,18 +160,22 @@
   },
   "devDependencies": {
     "@openapitools/openapi-generator-cli": "^2.4.26",
-    "@types/collections": "^5.1.2",
+    "@swc/core": "^1.2.144",
+    "@swc/helpers": "^0.3.3",
     "@types/express": "^4.17.13",
     "@types/mocha": "^9.1.0",
     "@types/node": "^17.0.16",
     "@types/node-fetch": "^3.0.3",
+    "@types/portscanner": "^2.1.1",
     "aegir": "^36.1.3",
     "check-aegir-project": "^1.0.3",
-    "collections": "^5.1.13",
     "dotenvrc": "^1.0.1",
     "mock-ipfs-pinning-service": "^0.4.0",
     "npm-run-all": "^4.1.5",
     "openapi-typescript": "^5.1.1",
+    "portscanner": "^2.2.0",
+    "regenerator-runtime": "^0.13.9",
+    "ts-node": "^10.5.0",
     "tsc-silent": "^1.2.1"
   },
   "exports": {
