+++ test/isomorphic-tests/configuration.ts
@@ -1,0 +1,13 @@
+/* eslint-env browser, node, mocha */
+
+import { expect } from 'aegir/utils/chai'
+import { Configuration, ConfigurationParameters } from '../../src'
+
+export default async (setup: () => Promise<unknown>) => {
+  describe('Configuration', () => {
+    it('Can be instantiated', () => {
+      const configuration: ConfigurationParameters = {}
+      expect(() => new Configuration(configuration)).not.to.throw()
+    })
+  })
+}
